feat(release): v0.4.0 — auto-sync hardening and credential pipeline fixes
