+++ .grype.yaml
@@ -21,3 +21,8 @@ ignore:
 
   # zlib 1.3.1-r2: fix is in alpine 3.23 base; resolved by base image pin
   - vulnerability: CVE-2026-27171
+
+  # ip-address 10.1.0 (GHSA-v2v4-37r5-5v8g, Medium): bundled inside express-rate-limit
+  # in the API image and inside npm's internal node_modules in the worker image.
+  # Severity is Medium — non-blocking under --fail-on high. Track for next dep refresh.
+  # - vulnerability: GHSA-v2v4-37r5-5v8g
