+++ apps/api/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@tokentimer/api",
-  "version": "0.3.2",
+  "version": "0.4.0",
   "private": true,
   "description": "TokenTimer API Service",
   "license": "BUSL-1.1",
@@ -14,11 +14,11 @@
     "lint:fix": "eslint . --fix"
   },
   "dependencies": {
-    "@aws-sdk/client-acm": "3.990.0",
-    "@aws-sdk/client-iam": "3.990.0",
-    "@aws-sdk/client-secrets-manager": "3.990.0",
-    "@aws-sdk/client-sts": "3.990.0",
-    "axios": "1.15.0",
+    "@aws-sdk/client-acm": "3.1045.0",
+    "@aws-sdk/client-iam": "3.1045.0",
+    "@aws-sdk/client-secrets-manager": "3.1045.0",
+    "@aws-sdk/client-sts": "3.1045.0",
+    "axios": "1.16.0",
     "bcryptjs": "3.0.3",
     "connect-pg-simple": "10.0.0",
     "cookie-parser": "1.4.7",
@@ -26,7 +26,7 @@
     "csrf-csrf": "3.2.2",
     "dotenv": "16.6.1",
     "express": "5.2.1",
-    "express-rate-limit": "8.3.1",
+    "express-rate-limit": "8.5.1",
     "express-session": "1.19.0",
     "express-slow-down": "3.0.1",
     "express-validator": "7.3.1",
