+++ apps/dashboard/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@tokentimer/dashboard",
-  "version": "0.3.2",
+  "version": "0.4.0",
   "private": true,
   "license": "BUSL-1.1",
   "scripts": {
@@ -22,7 +22,7 @@
     "@chakra-ui/react": "2.10.9",
     "@emotion/react": "11.14.0",
     "@emotion/styled": "11.14.1",
-    "axios": "1.15.0",
+    "axios": "1.16.0",
     "framer-motion": "11.18.2",
     "js-yaml": "4.1.1",
     "lucide-react": "0.553.0",
