+++ apps/worker/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@tokentimer/worker",
-  "version": "0.3.1",
+  "version": "0.4.0",
   "private": true,
   "license": "BUSL-1.1",
   "type": "module",
@@ -13,7 +13,7 @@
     "lint:fix": "eslint . --fix"
   },
   "dependencies": {
-    "axios": "1.15.0",
+    "axios": "1.16.0",
     "nodemailer": "8.0.5",
     "pg": "8.18.0",
     "prom-client": "15.1.3",
