+++ contracts.manifest.json
@@ -2,7 +2,7 @@
   "manifestVersion": 1,
   "repo": "tokentimer-core",
   "contractsPackage": "@tokentimer/contracts",
-  "version": "0.3.2",
+  "version": "0.4.0",
   "namespaces": [
     {
       "name": "queue",
