+++ deploy/compose/Dockerfile.dashboard
@@ -5,7 +5,7 @@ FROM node:22.22.2-alpine3.23 AS builder
 RUN apk update && apk upgrade --no-cache
 
 # Enable pnpm via corepack
-RUN corepack enable && corepack prepare pnpm@10.33.0 --activate
+RUN corepack enable && corepack prepare pnpm@10.33.4 --activate
 
 WORKDIR /app
