+++ deploy/compose/Dockerfile.worker
@@ -8,7 +8,7 @@ RUN apk update && apk upgrade --no-cache
 RUN apk add --no-cache curl
 
 # Enable pnpm via corepack
-RUN corepack enable && corepack prepare pnpm@10.33.0 --activate
+RUN corepack enable && corepack prepare pnpm@10.33.4 --activate
 
 # Create non-root user (UID/GID 1000 is taken by the built-in node user)
 RUN addgroup -g 1001 -S tokentimer && adduser -u 1001 -S tokentimer -G tokentimer
